Add documentation page.
This removes the need for the extra docs.secretchronicles.de subdomain.
Note that compared to the old site, this site does neither contain
the C++ docs, nor the HEAD documentation. Reason for this is that
people needing one of those will have the source tree checked out
anyway and can thus just build these docs themselves by issueing

  $ rake docs

in the tsc/ subdirectory of the sourcetree, which will also give
them the most recent version available.

Also note that the /docs directory linked to on the new page is
mapped with an “Alias” directive in the Apache configuration on
the server, thus it will empty if you run the site locally, but
it will show up on the server.